// src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Book = {
  __typename?: 'Book';
  id: Scalars['String'];
  name?: Maybe<Scalars['String']>;
};

export type Query = {
  __typename?: 'Query';
  hello?: Maybe<User>;
};


export type QueryHelloArgs = {
  id: Scalars['ID'];
};

export type User = {
  __typename?: 'User';
  book?: Maybe<Array<Maybe<Book>>>;
  id: Scalars['String'];
  name?: Maybe<Scalars['Boolean']>;
};